refactor: move deployment-bundle related files in their own directory
