+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
 	"editor.defaultFormatter": "esbenp.prettier-vscode",
 	"cSpell.words": [
+		"Abortable",
 		"cfetch",
 		"chatgpt",
 		"clipboardy",
@@ -21,6 +22,7 @@
 		"Miniflare",
 		"mockpm",
 		"mrbbot",
+		"mtls",
 		"outdir",
 		"outfile",
 		"pgrep",
@@ -41,6 +43,7 @@
 		"weakset",
 		"webassemblymemory",
 		"websockets",
+		"workerd",
 		"xxhash",
 		"workerd",
 		"zjcompt"
