+++ packages/wrangler/src/create-worker-preview.ts
@@ -1,11 +1,15 @@
 import { URL } from "node:url";
 import { fetch } from "undici";
 import { fetchResult } from "./cfetch";
-import { createWorkerUploadForm } from "./create-worker-upload-form";
+import { createWorkerUploadForm } from "./deployment-bundle/create-worker-upload-form";
 import { logger } from "./logger";
 import { parseJSON } from "./parse";
 import { getAccessToken } from "./user/access";
-import type { CfAccount, CfWorkerContext, CfWorkerInit } from "./worker";
+import type {
+	CfAccount,
+	CfWorkerContext,
+	CfWorkerInit,
+} from "./deployment-bundle/worker";
 import type { HeadersInit } from "undici";
 
 /**
