+++ packages/wrangler/src/dev.tsx
@@ -6,18 +6,18 @@ import getPort from "get-port";
 import { render } from "ink";
 import React from "react";
 import { findWranglerToml, printBindings, readConfig } from "./config";
+import { getEntry } from "./deployment-bundle/entry";
+import { identifyD1BindingsAsBeta } from "./deployment-bundle/worker";
 import Dev from "./dev/dev";
 import { getVarsForDev } from "./dev/dev-vars";
 import { getLocalPersistencePath } from "./dev/get-local-persistence-path";
 
 import { startDevServer } from "./dev/start-server";
-import { getEntry } from "./entry";
 import { logger } from "./logger";
 import * as metrics from "./metrics";
 import { getAssetPaths, getSiteAssetPaths } from "./sites";
 import { getAccountFromCache, loginOrRefreshIfRequired } from "./user";
 import { collectKeyValues } from "./utils/collectKeyValues";
-import { identifyD1BindingsAsBeta } from "./worker";
 import { getHostFromRoute, getZoneForRoute, getZoneIdFromHost } from "./zones";
 import {
 	DEFAULT_INSPECTOR_PORT,
@@ -30,9 +30,9 @@ import {
 } from "./index";
 import type { Config, Environment } from "./config";
 import type { Route, Rule } from "./config/environment";
+import type { CfWorkerInit, CfModule } from "./deployment-bundle/worker";
 import type { LoggerLevel } from "./logger";
 import type { EnablePagesAssetsServiceBindingOptions } from "./miniflare-cli/types";
-import type { CfWorkerInit, CfModule } from "./worker";
 import type {
 	CommonYargsArgv,
 	StrictYargsOptionsToInterface,
